% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{benchmarks}
\alias{benchmarks}
\title{Get a list of benchmarks}
\usage{
benchmarks(
  run_id = NULL,
  batch_id = NULL,
  simplifyVector = TRUE,
  flatten = TRUE,
  ...
)
}
\arguments{
\item{run_id}{the run_id to get benchmarks for (default: \code{NULL}, list all)}

\item{batch_id}{the batch_id to get benchmarks for (default: \code{NULL}, list all)}

\item{simplifyVector}{coerce JSON arrays containing only primitives into an atomic vector}

\item{flatten}{automatically \code{\link[jsonlite:flatten]{flatten()}} nested data frames into a single non-nested data frame}

\item{...}{arguments passed on to class specific \code{print} methods}
}
\value{
the response
}
\description{
Get a list of benchmarks
}
