% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare runs, benchmarks, batches, or commits}
\usage{
compare(
  type = c("runs", "benchmarks", "batches", "commits"),
  baseline,
  contender,
  ...
)
}
\arguments{
\item{type}{the type of comparison to make (one of: runs, benchmarks, batches, commits)}

\item{baseline}{the baseline sha of the entity to compare}

\item{contender}{the contender sha of the entity to compare}

\item{...}{
  Arguments passed on to \code{\link[httr2:resp_body_raw]{httr2::resp_body_json}}
  \describe{
    \item{\code{resp}}{A response object.}
    \item{\code{check_type}}{Check that response has expected content type? Set to
\code{FALSE} to suppress the automated check}
    \item{\code{simplifyVector}}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}
  }}
}
\value{
the JSON response
}
\description{
Compare runs, benchmarks, batches, or commits
}
