% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_results}
\alias{compare_results}
\title{Compare benchmark results}
\usage{
compare_results(
  baseline,
  contender,
  zscore_threshold = NULL,
  pairwise_percent_threshold = NULL,
  ...
)
}
\arguments{
\item{baseline}{the ID of the baseline result to compare}

\item{contender}{the ID of the contender result to compare}

\item{zscore_threshold}{the zscore threshold to mark regressions and improvements.
Default is defined at the Conbench api level.}

\item{pairwise_percent_threshold}{the pairwise_percent_threshold to mark regressions and improvements.
Default is defined at the Conbench api level.}

\item{...}{
  Arguments passed on to \code{\link[httr2:resp_body_raw]{httr2::resp_body_json}}
  \describe{
    \item{\code{resp}}{A response object.}
    \item{\code{check_type}}{Check that response has expected content type? Set to
\code{FALSE} to suppress the automated check}
    \item{\code{simplifyVector}}{Should JSON arrays containing only primitives (i.e.
booleans, numbers, and strings) be caused to atomic vectors?}
  }}
}
\value{
the JSON response
}
\description{
Compare benchmark results
}
