% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_runs}
\alias{compare_runs}
\title{Compare runs}
\usage{
compare_runs(
  baseline,
  contender,
  zscore_threshold = NULL,
  pairwise_percent_threshold = NULL
)
}
\arguments{
\item{baseline}{the ID of the baseline run to compare}

\item{contender}{the ID of the contender run to compare}

\item{zscore_threshold}{the zscore threshold to mark regressions and improvements.
Default is defined at the Conbench api level.}

\item{pairwise_percent_threshold}{the pairwise_percent_threshold to mark regressions and improvements.
Default is defined at the Conbench api level.}
}
\value{
a tibble of run comparisons
}
\description{
Compare runs
}
