% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{conbench_perform}
\alias{conbench_perform}
\title{A wrapper around \code{req_perform()} with Conbench auth}
\usage{
conbench_perform(data, ...)
}
\arguments{
\item{data}{the request to perform}

\item{...}{arguments passed to \code{req_perform()}}
}
\value{
the response
}
\description{
Takes all the same arguments as \code{req_perform()} as well as a \code{dot_conbench}.
It will add the cookie if it exists already, or it will attempt to get a new
cookie and run the request. Should be used at the end of an \code{httr2} pipeline.
}
