% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conbench-request.R
\name{conbench_request}
\alias{conbench_request}
\title{A wrapper around \code{request()} with Conbench URL}
\usage{
conbench_request(base_url = NULL)
}
\arguments{
\item{base_url}{the base URL to use}
}
\value{
the request
}
\description{
Takes all the same arguments as \code{request()}. Should be used at the beginning
of an \code{httr2} pipeline. By default, the conbench URL will be read from the
.conbench file in the current working directory.
}
