% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conbenchcoms-package.R
\docType{package}
\name{conbenchcoms-package}
\alias{conbenchcoms}
\alias{conbenchcoms-package}
\title{conbenchcoms: An API wrapper for conbench communications}
\description{
An R interface to the conbench API.
}
\author{
\strong{Maintainer}: Sam Albers \email{sam@voltrondata.com} (\href{https://orcid.org/0000-0002-9270-7884}{ORCID})

Authors:
\itemize{
  \item Jonathan Keane \email{jon@voltrondata.com} (\href{https://orcid.org/0000-0001-7087-9776}{ORCID}) [contributor]
  \item Edward Visel \email{edward@voltrondata.com} (\href{https://orcid.org/0000-0002-2811-6254}{ORCID}) [contributor]
  \item Austin Dickey \email{austin@voltrondata.com} [contributor]
}

Other contributors:
\itemize{
  \item Voltron Data [copyright holder, funder]
}

}
\keyword{internal}
