% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardware.R
\name{hardware}
\alias{hardware}
\title{Get hardware}
\usage{
hardware(id = NULL, simplifyVector = TRUE, flatten = TRUE, ...)
}
\arguments{
\item{id}{the id of one specific hardware to get data for.}

\item{simplifyVector}{coerce JSON arrays containing only primitives into an atomic vector}

\item{flatten}{automatically \code{\link[jsonlite:flatten]{flatten()}} nested data frames into a single non-nested data frame}

\item{...}{arguments passed on to class specific \code{print} methods}
}
\value{
the response
}
\description{
Get hardware
}
