% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{history}
\alias{history}
\title{Get history for a benchmark}
\usage{
history(benchmark_id)
}
\arguments{
\item{benchmark_id}{the ID of a benchmark result to get the history for}
}
\value{
a tibble of history
}
\description{
Get history for a benchmark
}
