% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Get a list of benchmark info}
\usage{
info(id = NULL, simplifyVector = TRUE, flatten = TRUE, ...)
}
\arguments{
\item{id}{the info id for benchmark info}

\item{simplifyVector}{coerce JSON arrays containing only primitives into an atomic vector}

\item{flatten}{automatically \code{\link[jsonlite:flatten]{flatten()}} nested data frames into a single non-nested data frame}

\item{...}{arguments passed on to class specific \code{print} methods}
}
\value{
the response
}
\description{
Get a list of benchmark info
}
