% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url_cleaner}
\alias{url_cleaner}
\title{Clean up an API URL into one that can be linked}
\usage{
url_cleaner(url)
}
\arguments{
\item{url}{the API URL (e.g. "http://conbench.dev/api/foo")}
}
\value{
c leaned, clickable URL (e.g. "https://conbench.dev/foo")
}
\description{
Clean up an API URL into one that can be linked
}
