% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-results.R
\name{benchmark_results}
\alias{benchmark_results}
\alias{benchmarks}
\title{Get a list of benchmark results}
\usage{
benchmark_results(
  run_id = NULL,
  run_reason = NULL,
  earliest_timestamp = NULL,
  latest_timestamp = NULL,
  ...
)

benchmarks(
  run_id = NULL,
  batch_id = NULL,
  run_reason = NULL,
  simplifyVector = TRUE,
  flatten = TRUE,
  ...
)
}
\arguments{
\item{run_id}{the run_id to get benchmarks for (default: \code{NULL}, list all)}

\item{run_reason}{a string to specify the run reason (default: \code{NULL}, list all)}

\item{earliest_timestamp}{the earliest benchmark result timestamp (default: \code{NULL}, go back as far as possible)}

\item{latest_timestamp}{the latest benchmark result timestamp (default: \code{NULL}, go up to the current time)}

\item{...}{arguments passed on to class specific \code{print} methods}

\item{batch_id}{deprecated}

\item{simplifyVector}{coerce JSON arrays containing only primitives into an atomic vector}

\item{flatten}{automatically \code{\link[jsonlite:flatten]{flatten()}} nested data frames into a single non-nested data frame}
}
\value{
a tibble of benchmark results
}
\description{
Get a list of benchmark results

\code{benchmarks} is deprecated method of getting benchmark results and will be
retired in a future release. Please use \code{benchmark_results} instead.
}
