% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{history}
\alias{history}
\title{Get history for a benchmark}
\usage{
history(benchmark_id, simplifyVector = TRUE, flatten = TRUE, ...)
}
\arguments{
\item{benchmark_id}{the hash of a benchmark to get the history for}

\item{simplifyVector}{coerce JSON arrays containing only primitives into an atomic vector}

\item{flatten}{automatically \code{\link[jsonlite:flatten]{flatten()}} nested data frames into a single non-nested data frame}

\item{...}{arguments passed on to class specific \code{print} methods}
}
\value{
the response
}
\description{
Get history for a benchmark
}
